+++ src/wasm/types.rs
@@ -7,6 +7,11 @@ const _: &'static str = r#"
 export type NomadIdentifier = string;
 export type NameOrDomain = number | string;
 
+export interface AppConfig {
+  displayName: string;
+  nativeTokenSymbol: string;
+}
+
 export interface NomadLocator {
   domain: number;
   id: NomadIdentifier;
@@ -98,6 +103,7 @@ export interface NetworkSpecs {
   blockTime: number | string;
   supports1559: boolean;
   confirmations: number | string;
+  blockExplorer: string;
 }
 
 export interface CustomTokenSpecifier {
@@ -134,11 +140,15 @@ export interface NomadConfig {
   core: Record<string, CoreContracts>;
   bridge: Record<string, BridgeContracts>;
   agent: Record<string, AgentConfig>;
+  bridgeGui: Record<string, AppConfig>;
 }
 "#;
 
 #[wasm_bindgen]
 extern "C" {
+    #[wasm_bindgen(typescript_type = "AddConfig")]
+    pub type AppConfig;
+
     #[wasm_bindgen(typescript_type = "NomadLocator")]
     pub type NomadLocator;